[WIP] Start work on code navigation for QObjects.
Allows for QObjects to navigate to the code location, where they were
created. Not (only) the header file or the class's constructor, but
where the constructor was called from.

This is done by generating a backtrace from
Probe::objectFullyConstructed, which is called from the QObject
constructor. From there we can step up to the frame above the object's
constructor and get the source location of that frame.

Current state:
- We know three ways of getting the backtrace: Sergio's backtrace
  tools, libunwind directly, GammaRay's getBacktrace() function. Each
  with it's own benefits. Currently implemented is backtrace-utils,
  but that's easy to change.
- We need a way to interpret the debug table in order to get the
  source location out of the frame address. For that we know two
  possibilities: Use the commandline tool addr2line directly or
  through backtrace-utils or use libbacktrace. None of it is currently
  implemented.
- GUI implementation is anywhere. Some stuff is done, but we need to
  check it against quickinspector's implementation and complete and
  fix it.